+++ app/src/main/java/network/minter/bipwallet/internal/di/InjectorsModule.java
@@ -52,7 +52,7 @@ import network.minter.bipwallet.internal.di.annotations.ServiceScope;
 import network.minter.bipwallet.security.SecurityModule;
 import network.minter.bipwallet.security.ui.PinEnterActivity;
 import network.minter.bipwallet.security.ui.PinValidationDialog;
-import network.minter.bipwallet.services.livebalance.LiveBalanceService;
+import network.minter.bipwallet.services.livebalance.RTMService;
 import network.minter.bipwallet.settings.ui.PasswordChangeMigrationActivity;
 import network.minter.bipwallet.tx.TransactionsModule;
 import network.minter.bipwallet.tx.ui.ExternalTransactionActivity;
@@ -128,7 +128,7 @@ public interface InjectorsModule {
 
     @ContributesAndroidInjector
     @ServiceScope
-    LiveBalanceService balanceUpdateService();
+    RTMService balanceUpdateService();
 
     @ContributesAndroidInjector
     @ActivityScope