+++ app/src/main/java/network/minter/bipwallet/internal/helpers/PrefKeys.java
@@ -34,4 +34,5 @@ public final class PrefKeys {
     public static final String ENABLE_SOUNDS = "pref_enable_sounds";
     public static final String ENABLE_PIN_CODE = "pref_sec_enable_pin";
     public static final String ENABLE_FP = "pref_sec_enable_fp";
+    public static final String LAST_BLOCK_TIME = "last_block_time";
 }