+++ app/src/main/java/network/minter/bipwallet/tx/views/ExternalTransactionPresenter.java
@@ -51,6 +51,7 @@ import network.minter.blockchain.models.operational.TxCoinBuy;
 import network.minter.blockchain.models.operational.TxCoinSell;
 import network.minter.blockchain.models.operational.TxCoinSellAll;
 import network.minter.blockchain.models.operational.TxCreateCoin;
+import network.minter.blockchain.models.operational.TxCreateMultisigAddress;
 import network.minter.blockchain.models.operational.TxDeclareCandidacy;
 import network.minter.blockchain.models.operational.TxDelegate;
 import network.minter.blockchain.models.operational.TxEditCandidate;
@@ -60,6 +61,7 @@ import network.minter.blockchain.models.operational.TxSendCoin;
 import network.minter.blockchain.models.operational.TxSetCandidateOffline;
 import network.minter.blockchain.models.operational.TxSetCandidateOnline;
 import network.minter.blockchain.models.operational.TxUnbound;
+import network.minter.blockchain.utils.Base64UrlSafe;
 import network.minter.core.MinterSDK;
 import network.minter.core.crypto.BytesData;
 import network.minter.core.crypto.MinterAddress;
@@ -107,6 +109,7 @@ public class ExternalTransactionPresenter extends MvpBasePresenter<ExternalTrans
         super.attachView(view);
         accountStorage.update();
         mFrom = secretStorage.getAddresses().get(0);
+        getViewState().setOnCancelListener(this::onCancel);
     }
 
     @Override
@@ -121,20 +124,43 @@ public class ExternalTransactionPresenter extends MvpBasePresenter<ExternalTrans
                 return;
             }
 
-            String rawPass;
-            try {
-                rawPass = params.getString("p", null);
-                if (rawPass == null) {
-                    mCheckPassword = null;
-                } else {
-                    mCheckPassword = RLPBoxed.decodeString(hexStringToChars(rawPass));
-                }
-            } catch (Throwable t) {
-                Timber.w(t, "Unable to decode check password");
-                mCheckPassword = null;
+
+            String hash = null;
+            // this value represents that we are parsing old-style deeplinks
+            boolean support = true;
+
+            if (!params.containsKey("d") && !params.containsKey("data")) {
+                getViewState().disableAll();
+                showTxErrorDialog("Unable to parse deeplink: No transaction data passed");
+                return;
+            } else if (params.containsKey("d")) {
+                hash = params.getString("d", null);
+            } else if (params.containsKey("data")) {
+                support = false;
+                hash = new BytesData(
+                        Base64UrlSafe.decode(params.getString("data", null).getBytes())
+                ).toHexString();
             }
 
-            final String hash = params.getString("d", null);
+            if (params.containsKey("p")) {
+                try {
+                    String rawPass = params.getString("p", null);
+                    if (rawPass == null) {
+                        mCheckPassword = null;
+                    } else {
+                        if (support) {
+                            mCheckPassword = RLPBoxed.decodeString(hexStringToChars(rawPass));
+                        } else {
+                            mCheckPassword = Base64UrlSafe.decodeString(rawPass);
+                        }
+
+                    }
+                } catch (Throwable t) {
+                    Timber.w(t, "Unable to decode check password");
+                    mCheckPassword = null;
+                }
+            }
+
             try {
                 mExtTx = DeepLinkHelper.parseRawTransaction(hash);
                 if (!validateTx()) {
@@ -235,7 +261,9 @@ public class ExternalTransactionPresenter extends MvpBasePresenter<ExternalTrans
     private boolean validateTx() {
         if (mExtTx.getType() == OperationType.RedeemCheck) {
             TxRedeemCheck d = mExtTx.getData();
-            if (d.getProof().size() == 0 && mCheckPassword == null) {
+            if (mCheckPassword != null) {
+                d.setProof(CheckTransaction.makeProof(mFrom, mCheckPassword));
+            } else if (d.getProof().size() == 0 && mCheckPassword == null) {
                 getViewState().disableAll();
 
                 getViewState().startDialog(false, ctx -> new WalletConfirmDialog.Builder(ctx, "Unable to scan transaction")
@@ -246,8 +274,21 @@ public class ExternalTransactionPresenter extends MvpBasePresenter<ExternalTrans
                         })
                         .create());
                 return false;
-            } else if (d.getProof().size() == 0 && mCheckPassword != null) {
-                d.setProof(CheckTransaction.makeProof(mFrom, mCheckPassword));
+            }
+        } else if (mExtTx.getType() == OperationType.SellAllCoins) {
+            final TxCoinSellAll data = mExtTx.getData(TxCoinSellAll.class);
+            Optional<CoinAccount> acc = accountStorage.getData().findAccountByCoin(data.getCoinToSell());
+            if (acc.isEmpty()) {
+                getViewState().disableAll();
+
+                getViewState().startDialog(false, ctx -> new WalletConfirmDialog.Builder(ctx, "Unable to send transaction")
+                        .setText("You are trying to sell all %s for %s, but your %s balance is 0", data.getCoinToSell(), data.getCoinToBuy(), data.getCoinToSell())
+                        .setPositiveAction(R.string.btn_close, (_d, _w) -> {
+                            _d.dismiss();
+                            getViewState().finishCancel();
+                        })
+                        .create());
+                return false;
             }
         }
 
@@ -305,7 +346,6 @@ public class ExternalTransactionPresenter extends MvpBasePresenter<ExternalTrans
         getViewState().setSecondVisible(View.VISIBLE);
 
         getViewState().setOnConfirmListener(this::onSubmit);
-        getViewState().setOnCancelListener(this::onCancel);
 
         switch (tx.getType()) {
             case SendCoin: {
@@ -332,6 +372,8 @@ public class ExternalTransactionPresenter extends MvpBasePresenter<ExternalTrans
                 Optional<CoinAccount> acc = accountStorage.getData().findAccountByCoin(data.getCoinToSell());
                 if (acc.isPresent()) {
                     getViewState().setFirstValue(String.format("%s %s", bdHuman(acc.get().getBalance()), data.getCoinToSell()));
+                } else {
+                    getViewState().setFirstValue(String.format("%s %s", bdHuman(0), data.getCoinToSell()));
                 }
 
                 getViewState().setSecondLabel("For");
@@ -424,6 +466,22 @@ public class ExternalTransactionPresenter extends MvpBasePresenter<ExternalTrans
             }
             break;
 
+            case CreateMultisigAddress: {
+                final TxCreateMultisigAddress data = tx.getData();
+                getViewState().setFirstLabel("You're creating MultiSig address");
+                StringBuilder sb = new StringBuilder();
+                for (int i = 0; i < data.getAddresses().size(); i++) {
+                    MinterAddress address = data.getAddresses().get(i);
+                    Long weight = data.getWeights().get(i);
+                    sb.append("Weight: ").append(weight).append(" - ").append(address.toShortString()).append('\n');
+                }
+                getViewState().setFirstValue(sb.toString());
+                getViewState().setSecondLabel("Threshold");
+                getViewState().setSecondValue(String.valueOf(data.getThreshold()));
+                getViewState().setSecondVisible(View.VISIBLE);
+            }
+            break;
+
             case EditCandidate: {
                 final TxEditCandidate data = tx.getData(TxEditCandidate.class);
                 getViewState().setFirstLabel("You're editing candidate");
@@ -443,7 +501,6 @@ public class ExternalTransactionPresenter extends MvpBasePresenter<ExternalTrans
                         .setPositiveAction(R.string.btn_close)
                         .create());
             }
-
         }
     }