+++ RELEASE_NOTES.md
@@ -1,5 +1,13 @@
 # Release Notes
 
+## 1.14.0
+ - Added support new deeplinks protocol
+ - Added information about last updated block
+ - Fixed cancel button at external transaction activity after app recreated
+
+## 1.13.0
+ - Texasnet release
+
 ## 1.10.0
  - Now app supports only android starting version 5.0 (api 21)
  - Added check redeeming ability